ScoreAPI/requirements.txt
fastapi
uvicorn
kafka-python
psycopg2-binary
python-jose
passlib[bcrypt]
jinja2
